Fix: account for hiding monsters in mdisplacem (from xNetHack).
From commit 57ae2fa263:

'Was getting sanity check failures where a monster was concealing itself
under nothing. I believe it's due to the mvm displacement logic, where
the two monsters just are placed in each other's spots without checking
for if one of them is currently hiding.

Make them account for this. The monster initiating the displacement will
call hideunder() if it's capable of that and hide if it chooses to; the
monster being displaced is startled enough not to immediately hide, and
always becomes revealed.'