Tick drain life up to level 3 (from xNetHack).
Commit 511ff257a:

'This is another Fourk change. Drain life is a bit of an odd duck among
attack spells in that it's not used all that often for actually
*attacking* things, compared to force bolt and magic missile. Other uses
I know of include:
- Ascension tricks like draining the Riders so that they can be tamed.
- As a noncombat spell, lowering a +6 weapon to +5 or +4 armor to +3 so
  that it can be reenchanted.
- Deliberately draining the player's own experience level as part of
  drain-for-gain. Strategies like that were nerfed in 3.6, so I doubt
  this happens *as much*, but it is still certainly a thing to e.g.
  drain yourself to XL1 so you can buy lots of cheap protection from a
  priest, then use blessed restore ability to get the levels back.

As an attack spell, it's arguably weaker than magic missile, arguably
slightly stronger. But this level boost will make it slightly more
difficult to use as a utility spell.'