Make Wintercloak a unique monster.
Since she is the proper guardian of Sokoban, this seems fitting. I added a few small buffs so she's a little stronger than the typical white dragon.

I could have gone the extra step of creating an achievement for killing her, but since we already have acquiring the Sokoban prize as a big event it felt like too much.