Lessen harshness of enchanting antimagic items.
We'll be a bit more lenient, so that there is a reasonable chance of enchanting these items up at least a little bit at first. I tried to match the behavior of ring enchantment, so that there is a low chance of failure at low enchantment, but it scales at you get higher.

For enchant armor:
+0	 0%
+1	14%
+2	29%
+3	43%
+4	57%

For enchant weapon:
Wep	Chance of failure
+0	 0%
+1	 8%
+2	15%
+3	23%
+4	31%
+5	38%
+6	46%
+7	54%
+8	62%
+9	69%
+10	77%
+11	85%
+12	92%