Fix and refine artifact to-hit penalties.
I realized there were a couple issues with this. If you were using an artifact with a hated monster type, like Sunsword, the penalty would only apply vs undead. I fixed this so that the -15 to-hit penalty applies to everything upfront, but banes get a generous +25 to-hit bonus when used vs their banes. So Sunsword vs undead is effectively +10 to-hit, making it much better and situational.