Fix: dinosaur gen for cavemen was broken.
I found this bug from copying the method SpliceHack used for enabling dinos for something else (attempting to make winter wolves domestic for valkyries). But unfortunately, when you set a flag in u_init in the beginning of the game, it doesn't carry over when the game is saved and restored. I was starting to wonder about this because after many hours of playing cavemen, I never had seen the mythic velociraptors or t-rexes... Instead, the dinos will now have standard generation, but the role restriction will be handled in makemon.c as opposed to manipulating flags. I'm not sure if this is 100% the best route, there might be a few loose strings.