Fix: Gnomish potions of oil could initialize with +1 spe. Fixes #67.
This solves the problem of stacks of water not merging. Turns out that the spe field had a one when I added this change for the gnomish race. When diluting this potion, the enchantment field would not be wiped so they would not merge. I messed around a little with the rknown field too, but it doesn't really matter for potions since they merge regardless.