Fix: vault guards were not entering vaults filled with grass.
I'm lucky I caught this bug purely by accident and teleporting into a vault. I waited many hundreds of turns but no guard came!  My only guess to the problem was the vault was filled with grass and that turned out to be the problem. Now, shops and vaults will be prevented from having grass spawn inside.