Lethe improvements.
Restored some of the power items that one could find here and made altars always guaranteed to appear in their respective levels (but they still may be cracked).

These levels need to have some decent rewards since they are 1) difficult and 2) possibly grindy.

I also cut down on the chances that most of them appear. Only leth-a, b, and z are guaranteed. All the others are 25-50% to appear, hopefully resulting in an average lethe branch of 3-4 levels.

I'm also going to properly separate the leth-a level to it's own branch that extends up from leth-b, so that will also shorten the length a bit. However, there is a guaranteed altar and magic lamp on that level, so I think players will still want to complete it.