Implement new dragon armor system: DSM replaced by scaled armor (from xNetHack/Fourk).
Ported straight from xNetHack's implementation, based mostly off of commit 00fec3777:

'This is an implementation of what is commonly termed the 'dtsund-DSM'
proposal (http://roguelikes.live/nh4-bugzilla/show_bug.cgi?id=629) by
dtsund and jonadab: rather than having dragon scale mail objects as
their own armor, create a system by which the player can make
'dragon-scaled' armor - a regular leather armor, chain mail, etc, with
dragon scales embedded into it, conferring the same extrinsic they give
normally. This implementation is based on NetHack Fourk's, but with
heavy adjustments because a lot of the core NetHack code is changed
there.

The main reason for replacing dragon scale mail with this new system is
that dragon scale mail is a supremely optimal choice of armor and makes
armor strategy much less interesting. To summarize, dragon scale mail
was ALL of the following simultaneously:
  a) lightweight
  b) higher in base AC than all other armors
  c) naturally erodeproof
  d) transparent to spellcasting
  e) available before the Castle and sometimes even the Quest, or from
     the first early wish (and it was the #1 most common early wish)
  f) a provider of a powerful extrinsic (reflection, magic resistance)
Once dragon scale mail became available, there was absolutely no reason
not to wear it. With it out of the picture, though, armor strategy
becomes relevant again past the first few dragons in the game, because
the player ultimately needs to decide on one of the base armors to wear,
and will have to continue considering the tradeoffs that come with it.

This also is a big step forward for the xNetHack philosophy of
encouraging found objects over wished ones. Previously, if you found a
really nice mithril chain mail or something, you would know you would
only keep it around until you could acquire dragon scale mail, and then
ditch it. Now the best option is to find a good color of dragon scales,
attach it to your armor, and keep using it.

That said, I don't want this to come across as a nerf for its own sake.
I plan on mitigating the weight factor by implementing another variant's
adjustments to body armor weight while it's being worn. This is also
why dragon scales still give 3 AC when worn as a cloak and give +3 AC to
any scaled armor: it seemed like a pointless nerf to drop the AC of
scales to zero (meaning you can do pretty well AC and extrinsic wise by
wearing scaled armor and scales of a different color over top, at the
cost of having zero MC), and scaled armor ought to derive _some_ benefit
from its impervious dragon scales other than the extrinsic.

Another difference from Fourk is that xNetHack scaled armor is not
automatically erodeproof. Mainly because it seemed more interesting to
keep around the possibility of the player's armor continuing to be
vulnerable to erosion past the midgame. I assume most players will
erodeproof their armor or put scales on armor made of a material that is
naturally erodeproof, anyway.

As in Fourk, I didn't bother implementing the part of the original
proposal where you can pry the scales back off armor with a tin opener,
because it seemed really weird flavor-wise and I didn't want to make an
occupation for spending multiple turns prying scales off.

As in Fourk, if you polyself and merge with your scaled armor, you get
it back after unpolymorphing, which is an intentional buff. Doesn't seem
right to have its scales unmeld from it and cost the player another
enchant scroll to put the scales back on.'