Any role that could start with a magic marker always starts with one.
This is kind of a backpedal on previous changes, but hear me out this serves two purposes: it makes the game easier for the weaker roles (and they desperately need that), and it cuts down on start-scumming. If I was going to enable them randomly like in vanilla, players will just scum for them anyway, so this saves everybody some time and server overhead.