Auto-identify items that have an observable effect when monsters wear them.
This was reported by loggers for the amulet of flying, but I found a few other places where we could benefit. This turned into a bit of refactoring, so the muse code became a bit more readable.

My current solution focuses more on witnessing the action of a monster wearing an item - I left the un-wearing as a separate exercise. Unwearing is quite a bit more rare anyway, so I don't think this will have much effect.