Endgame genocide effectiveness depends on the monster's proximity to the hero.
I think this is a much better solution to the endgame geno nerf I tried before and less random. Instead of a flat 20% chance of killing all the specified monster species on the level, we instead go by how close the monster is to the hero. If the monster is next to or within 3 squares - guarantee death. Otherwise, the chance drops as the distance increases. This way, if the hero is in a crunch with a nasty really close, they know they can depend on the scroll (ie: with a kraken trying to drown you).

This might even lead to a general change in the future to how this scroll works. Maybe blessed will destroy all on the level and uncursed will take this general behavior. We'll see...