Lower level players get an additional to-hit bonus.
Trying to balance out some of the to-hit nerfs a little. I don't really want to undo the strength based to-hit nerf, but this seems like a good alternative, since it builds on a previous idea.

While under level 5, players get an additional +1 to-hit bonus. So this means that for levels 1-2, players enjoy a +2 to-hit bonus, which then tapers to +1 for levels 3-5 - then hopefully they have figured out that works.