Fix: wizard quest nemesis was not getting bell of opening.
Reported by Arseniy. In a previous update to the Dark One's inventory, it turns
out that specifying an inventory in LUA discards any existing inventory, and
effectively deletes the Bell of Opening. I just moved the inventory adjustments
to makemon.c alongside other nemesis inventory adjustments.

The enchantments of the items were slightly changed - the cloak of magic
resistance is just +0 now, and the staff is now at least +1 to +3.