Cartomancers start with a playing card deck or a deck of fate.
After a little discussion in the #nerfhack IRC, this sounded like a really fun idea. We'll also see a lot more playtesting of these two items, which will hopefully result in bug fixes. The deck of fate could result in some pretty wacky/swingy starts, but it does feel in spirit with the role.