Non-magical alchemy is less likely to result in an alchemical explosion.
hothraxxa noticed that mixing potions of blood was frequently being subject to alchemical explosions. This change makes the chance of an explosion with non-magical ingredients 1 in 20.

I didn't really want to hard-code exceptions for blood because it there are multiple recipes for blood potions, some of which also involve fruit juice. This method seems the most elegant and hopefully brings a bit more sense to alchemy.