Add a status for Rabid. Fixes #54.
Rabid is a very nasty status, so it definitely deserves it's own place on the status bar. I found that there was an Iron status that was half implemented for when elves might be touching iron, but since I don't plan on ever using it I decided to replace it with Rabid.

I think I was able to fill in the rest of the gaps, however in win/X11/winstat.c, I had to replace the status version to fit it in. This will probably only apply for windows versions, but it might be worth looking into expanding the status columns to 7 lines.