Nerf vault guard farming.
This was originally brought to my attention by Arseniy, but it has been a known exploit in NetHack for a long time. Once a player is ready or has a pet capable of taking care of vault guards, it's possible to summon guards on demand with a cursed whistle and farm them infinitely. The death drops won't occur because the guard is usually in a wall, but they are capable of spawning with defensive or offensive items.

My solution to this was to prevent the spawn items and the death drops (in the event someone figures out how to farm them for those). I also buffed their stats slightly to compensate for the lack of items.

My hope for this kind of nerf is to encourage found items and natural exploration of the dungeon and discourage repetitive and boring farming.