Werefoo farming nerf; werefoo summon temporary spell-being breathren.
With the nerfing of other forms of farming, players may turn to more desperate measures like this. In the spirit of NerfHack, I'm happy to  be waiting with a big NOPE.

This is a also neat usage of the spell beings. It may turn out to be a blessing in disguise for players because the usual nightmare of being surrounded by were-summons is now only a temporary problem. The temporary summons also apply to players.