Revamp how piercers damage hard helmets.
After getting some feedback on the piercers destroying helmets, it did seem a tad unfair the helmet was totally destroyed. I changed it so that instead of outright destruction, the helmet is disenchanted -1 (or -d3 with negative luck). Once the helmet has negative enchantment, it can be totally destroyed.

I also incorporated some handling for the new glass crystal helmets, so they progressively crack instead of receiving enchantment hits.