Skewering tweaks.
Aosdict mentioned some interesting ideas in the #nerfhack IRC that would make skewering both more interesting and more balanced.

As it was, there were no real restrictions on when or how skewering could happen if the skill was met and there were no peaceful monsters in the way. However, aosdict thought that it would make more sense if the physics of the actual monsters was taken into account - for example, that unsolid monsters can always be skewered and that the player could not skewer through a normal monster unless it was a killing blow.

This made sense, so I followed it through to the logical and most NetHacky ends. I added amorphous monsters, molds, and blobs because they should be squishy enough to always skewer. I also added "kebabable" monsters - because that mechanic in Vanilla deserves a little extra spotlight and it actually makes sense here.

And finally, for monsters that are none of the above but still fleshy and can be skewered - they must be below 20% health.

To make it more accessible, I lowered the skill requirements for spear, trident, and polearm (with spetum) by 1 notch so that it's easier to use skewering. Hopefully it results in better and more concise fighting.