Being polymorphed into a wandering form will sometimes make you wander.
This is an idea that came to me when a player asked about good polyforms in NerfHack. I noticed that the only form that wandered and granted stunning was the invisible stalker form. However, many powerful forms do wander, notably: byakhees, ghoul mages, ghoul queens, gugs, revenants, migo warriors, and migo queens. This is not so much a polyform nerf as just trying to make game mechanics more consistent.

The vanilla NetHack behavior was that monsters that wander (with the M1_WANDER tag) have a 1 in 4 chance of randomly moving each turn. I made it 1 in 8 for polyd players so it's not *as* bad but will still have a very noticeable impact.