Racial and polyd vampires won't attempt fatal bites on monsters.
I think this subtle but cheeky check that dates back to SLASH'EM vampires is a bit too niche and dangerous and usually results in stupid deaths. I also removed the check for wielding a weapon, so even if you are bare-handed, you won't attempt a fatal bite (you still might punch a cockatrice and get stoned though).