Fix: null uwep access in Sting_effects (reported by hothraxxa).
Sting_effects is the function that determines how much "glow" an item or artifact has when it's warning of monsters nearby. Mayhem warns of undead, so it didn't fit into the existing assumptions about the items that confer warning. Since undead are so common and the player will be able to see them on-screen, I'm not going to handle the glow effects for Mayhem right now (or maybe ever). But we can at least patch the crash.