The Chromatic Dragon can berserk.
This nemesis doesn't have much for changes other than lower AC. My rationale is that part of this dragon is red (it's chromatic right?) so it still has the potential for a lot of explosive anger. The other reason we can buff it is that the caveman's guaranteed sacrifice gift, Skullcrusher, is now an aklys. So one could probably hide in a boulder fort with that and semi-safely kill it from range. Playtesting will reveal if this is too nasty.