Occasional critical hits for veteran monks.
Monks that have reached grand master in martial arts and level 21+ start getting a 1 in 20 chance for critical hits when fighting bare-handed. If you hit the 1, you inflict double damage. It's a nice cherry on top for monks that have put in the time and encourages reaching those higher levels.