Cleanup some things with aggravate monster timing out.
hothraxxa noticed that aggravate monster didn't seem to be timing out so I looked into it and didn't find any major issues, but a couple places where info could be more detailed. Now we'll show the timeout in wizmode for testing, and give the player a message when it wears off in-game.