Fix: Invocation spot was not aligned with vibrating square, rendering invocation impossible (reported by hothraxxa).
This is not a great fix, but it is fairly simple. For some reason the coordinates of (svi.inv_pos.x, svi.inv_pos.y) are being mixed up when the vibrating square is placed. This is a problem introduced with a fixed nkai-z.lua map, but since I plan to remove those maps this is an okay solution for the time being.