Fix: pokedex said quantum mechanics grant teleportation.
This is currently a bug in variants with the pokedex. The pokedex lookup for quantum mechanics falsely states that they grant the teleportation intrinsic. This is because they have the M1_TPORT flag, but when eaten they only grant speed or slowness. The other standard intrinsics are skipped.

This is a patch fix - if more monsters start showing up that violate the pokedex we'll implement a less fragile fix.