The High Priestess grants charisma and temp ESP instead of fixing alignment.
The alignment record bonus inadvertantly conflicts with my agenda to make artifact wishes very difficult for early cartomancers. The previous effect would set the players alignment abuse record to 0 (debatable if that should be allowed in the first place) and grants them 10 alignment points. This went way beyond the -1 that the deck of fate sets when applied.