New artifact: acidfall (from SpliceHack).
This is a long sword very similar to the other brands which deals double acid damage and grants extrinsic acid resistance. I borrowed a few things from Dirge, so this is now also a chaotic intelligent artifact. However, we didn't have a dark knight, so differs from Dirge in that it can generate randomly and doesn't require a same-race sacrifice to attain.

Unlike Splice's Acidfall, this weapon will inflict a lot of item damage, frequently corroding the items and armor it hits. Because acid resistance is rare among monsters, this weapon would be overpowered. Making it destructive and keyed to chaotics tones it down a bit.