New artifact: Glamdring (from EvilHack).
This is the only forged artifact I will port - but this one just felt like it would be a lot of fun to play with. I had to change the type to a standard long sword because we don't have object materials, but that's fine. I also gave this +10 damage vs all monsters, so it behaves similar to Angelslayer or Grimtooth in that way. It can generate randomly, can be wished for, and doesn't require forging.

Since its a bane, I gave it a minimum of +1 when it's received via #offer gift.