Move spellbooks of fireball and cone of cold to matter spell school; fire bolt to attack.
This is a slight balancing of attack spells - it makes wizards have to work a bit harder for these spells since they can't just spam force bolt and level up attack spells for these. Fire bolt was moved to attack because it fits with force bolt.

Another factor is that chain lighting was recently added to the attack spells, giving the ATTACK school 3 powerhouse spells to work with: force bolt, chain lightning, and finger of death.