Add phasing status.
Since we have potions of phasing to make this condition more common, a status line for this is a nice QoL update. I decided to replace the existing unused status for being "unconscious" since it seems unlikely it will ever get used. If you are unconscious, what good is a status?