Cockatrices only inflict slow stoning for the player and for monsters.
The bulk of this was ported from EvilHack, however, there were still many instances of instant petrification left over so I forged ahead and made the mechanic consistent for all other situations, including bumping into a stoning monster, touching a stoning corpse, etc.

This acts as BOTH a nerf and a buff. It really nerfs the ability to instakill monsters with the c-corpse - but it also prevents a lot of meaningless YASD's from common accidents. It's very easy to typo in the late game and accidentally eat a cockatrice corpse or touch one. You will still start the stoning process but you will at least have the chance to stop the death.

The nerf does seriously impact a few key parts of the game:
* the astral plane will be much more difficult without the "rubber chicken" blazing through the crowds.
* players will have one less tool to use against the wizard of yendor or other strong monsters.