You will always hit monsters who are holding you.
Inspired by a complaint from loggers who was lamenting a character that died while being by something but kept missing. The behavior should now match being swallowed, where you are always guaranteed to hit the engulfer.

This could probably be improved upon, maybe the chance should just be significantly higher, but this was simple and we'll get to see how it plays out.