Spread wishes throughout Gehennom via scrolls of wishing.
This is a spin on the xNetHack idea of spreading the wishes/love around the endgame instead of dumping them all at the castle. This hopefully makes for a more interesting and rewarding endgame.

Scrolls of wishing will appear at:
* The Castle
* Vlad's Tower
* The Wizard's Tower
* Moloch's Sanctum (replacing the previous magic marker there)

I used zap cards of wishing because:
A) they can only be used once. It is awkward to make wands unchargable or already charged without confusing some or all players. This skips the whole charges issue and presents the user with an easy choice.
B) I already had these setup from the cartomancer role, making this a super easy conversion.
C) These scrolls can also be zapped, making them friendly for illiterate conduct. If a non-cartomancer has maintained illiterate conduct, the scroll will make it obvious they can be zapped.

This change also has the player benefit of guaranteeing 4 wishes instead of the previous 3 from the castle WoW.