New monster: leper (l).
For context: this one came to me while flying back from America to Sweden on a 7.5 hour flight. I thought it would be interesting to have a monster with the disease threat (similar to evilhack's zombies), but that is pretty weak - definitely much slower and less dangerous than locusts. So now we have the leper - an additional l class monster. As a weird side effect, it inherits some of the stuff from the S_LEPRECHAUN class, so they generate sleeping and occasionally you will "miss wildly".

Fun fact: leprosy is one of humanity's oldest diseases.