mplayers can steal the amulet from the player (from EvilHack).
This is a nasty mechanic that seems to have originated from SpliceHack with many refinements from EvilHack. I'm adopting a reduced form where the monster can "merely" steal it, but won't actually try to sacrifice it. If there is player demand for that secondary mechanic, we'll have to revisit it but it is a pretty brutal form of instadeath. On the other hand, it only seems to have happened maybe once or twice in EvilHack's history. The fact that its so rare makes me even less inclined to port it though...

One major change: the frequency for mplayers stealing is 1 in 3 as opposed to the Wizard of Yendor's 1 in 20 chance.