Monsters carrying the Amulet of Yendor are incapable of teleportation.
This is an extended form of ffd359e03e from EvilHack, which restricted monsters from teleporting in the Sanctum. I haven't tested this, but it might be interesting if the Wizard of Yendor or other covetous monsters are not able to teleport once they get this. It could makes fights a little more fair and less tedious.