Increase the Constitution penalty for getting lifesaved (from xNetHack).
From commit 1f04e171f2:

'The existing penalty of -1 Con that can be restored right away is quite
mild. Granted, it's not a total non-problem since unicorn horns no
longer work to fix that, but it's still just a quaff or cast of restore
ability away from being cured. The penalty should stick around a bit
more.

This increases the Con penalty to -2, and additionally takes the same
bite out of your maximum Con attained, meaning it will take exercise or
gain ability to fix. I thought about permanently decreasing the maximum
possible value for it as well, but it turns out that value is
overwritten when a save is reloaded, and that's more trouble than I'm
willing to go to right now.'