Wands are used up even if they fail their wresting chance; much higher wresting odds (from EvilHack).
This is another small but pretty effective change. It's a nerf in the sense that players will have less wands sitting around that they can sell or recharge, but it also makes the process of wresting a lot less time-consuming and tedious. Because we aren't using wands of wishing, players don't have to worry about losing their WoW's to this change.