Allow covetous monsters to follow normal AI (from xNetHack).
I thought this was a really cool change with the demon lords, makes it so you can actually use some different tactics against them rather than just "go to the stairs".

From commit fdadff64:

'This adds a macro that attempts to disable covetous behavior in cases
where it pertains to warping and special movement behavior, without
affecting the other ramifications of the flag such as not allowing them
to be pacified. This included a special case in the nemesis dialogue
because apparently that code is reliant on the nemesis not being next to
you and then warping to you...

Downside: Covetous nonwarpers will not appear to go after the coveted
object at all. E.g. if it's sitting on the floor, or held by another
monster, they won't go there or attack the monster; they'll just keep
attacking the player. For quest nemeses, where neither of these
situations is likely to happen much, this is acceptable right now. I
think it would take a larger AI rewrite to allow nemeses to chase their
target without the ability to instantly arrive there.'