The Death card can result in polymorph if your luck is high.
My Tarot knowledge is pretty basic, but I know that the Death card isn't just a dark card of doom - it's also a card of change.

Here, if the player has at least 7 luck and avoids a 1 in 13 roll, they will avoid certain death and transform into something else for a while. Unchanging will also negate the transformation so you'll get the touch of death.

This should make the deck a little more user friendly, and it gives the early scumming cartomancers a saving grace because sometimes they'll hit the Moon (+7) luck and then Death and maybe survive.