Port the Wyrm Caves from SLASH'EM.
This will be the final branch I will import for NerfHack. With the removal of the Lethe levels, I thought something was lacking for the midgame. This is one of my favorite levels from SLASH'EM, it's simple but if offers the player some valuable loot for that stage: gems, dragon scales, and worm teeth.