New artifact: Mortality Dial (from SpliceHack).
This was one I also played with in HackEM, but here it's a little different. It's a morning star now that grants warning and regeneration when wielded. It also prevents all monster regeneration. Since it's not an executioner's mace, I bumped the damage bonus up to d12.

A lot of the stuff I added lately was chaotic, so here's one for the lawfuls.