Crystal plate mail updates (from EvilHack).
To make this armor a bit more enticing I ported most of the changes from Evil (but not all).
* CPM only weighs 150aum.
* Base price is 999
* Made of gemstone instead of glass, so it won't shatter from force bolts or wands of striking.
* Resistant to destruction or disintegration

I made another tier of armor for monsters so that monsters with -3 or lower only have a chance to get CPM. Previously, watch captains and vault guards could get these and I want to push these much deeper into the dungeon. I also lowered the probability of these quite a bit so they will be quite rare unless a monster drops them. I thought about adding the forging recipe, but because I want the forges to be primarily for weapons I left that out.

I also reverted the AC protection back to 7AC since the weight is so much less.