You cannot be fast while flying or levitating.
For a long while I've wondered how to address the fact that speed boots are S-tier items and there almost is no decision in whether to wear them or not in the endgame. To be clear, I don't think it's *bad* that they are a great item but I think there could be a bit more nuance in their effects.

This commit aims to put a crack in between two of the most powerful mechanics in the game: speed and flying. You will have to choose between the two now because the combination of speed with flight is too powerful. This is amplified by the fact that we offer vampires as a playable race, which start out with flying. I don't think it's an understatement that speed itself is probably the most important factor in the game. Having great speed acts as a damage multiplier and defense multiplier and just overall lets you make better decisions because you have more time(turns) to act.

One of my first ideas was to nerf speed somehow itself, but I couldn't come up with a good approach and it felt a little dirty to muck up a system which already seemed well done.

Part of my motivation for this is just the physics of it: you should need "boots on the ground" for speed boots to work (or any other item that grants speed). If you are in the air, it's difficult to see how you would have the traction necessary to propel yourself.

The other part of my motivation is to encourage different ascension builds and give vampires a little more of a challenge. Two roles that will notice the impact of this change more are vampire archeologists and vampire barbarbians who both can acquire intrinsic speed fairly early as part of the role abilities. Vampires in general will also have to contend with stunted speed potential. However, I plan to offer an option or two to 'ground' themselves in the future.