Stomping boots block flying.
This is a followup to the previous flying-vs-speed commit. Stomping boots should also require constant contact with the ground. I also wanted this change because it gives vampires a way to ground themselves and take advantage of speed if they so choose.

Stomping already did not function if the player was levitating so this is a logical extension of thatidea.