Disable bones files unless for wizmode testing.
I think disabling bones serves the core interests of NerfHack. On one hand, bones can serve as a way for players to get a leg-up on the game, using the items and drops from previous adventurers. On the other, it can force players to confront the same out-of-depth challenges that have killed one or more players, creating an annoying block point in the game.

But in Nerf, I want people to face a fresh dungeon each time. And if they win, I always want them to know they did it on their own without any external help. This also sets NerfHack apart from every other variant because no other variant has disabled bones AFAIK - and that is a good thing.