Mirrors offer reflection while carried.
This is a combination of xNetHack's commit 606a8e46e and YANI #4528 by ais523 (https://nethack-yanis.github.io/yanis/4528.html).

While carrying a mirror in open inventory, it will give you reflection. However, whenever the mirror reflects a ray, there is about a 50% chance it will be destroyed upon impact. I used this check: if 6d6 > 20 then the mirror is destroyed. This is not as accurate as I'd like, I'd rather the actual damage of the wand or ray source be evaluated, but I think this is fine. When a mirror is destroyed in this fashion, you do not take a luck penalty unless you were the originator of the ray.

For the other effects that are reflected - like floating eyes or medusa, the mirror is not destroyed. I'm not sure about this yet, but it seems fine since mirrors are fragile and susceptible to wands of striking and force bolt.