Fix: exploding drain life blast could cause is already detached error.
I believe the cause was the drain life ray traveling through monsters that were already killed by the blast, or the blast killing a monster already dead from the ray. Now the drain life spell is instead converted to a fireball type blast instead of being handled in the ray hitting code.