Tweak flying vs speed nerf.
These changes introduce a little more nuance to commit 9369b34d2 - 'You cannot be fast while flying or levitating.'

Completely eliminating any form of speed while flying is quite detrimental and probably a bit unbalanced. Fast speed grants the player a 33% increase in turns whereas Very Fast speed grants a 66% increase - both of which are an incredible boon to the player in all situations. I still am holding firm to the direction of the original nerf, but I think these changes allow for some lenience with regards to vampires and also cartomancers which start out a little slow.

The adjustments are as follows:
* Levitation blocks any form of speed from working.
* Flying allows for Fast speed but blocks Very Fast speed.

This middle ground preserves the challenge intended by the original change without making flying excessively punitive. It also ensures that characters with natural flight or those who heavily invest in speed as part of their build can still benefit, albeit to a lesser extent, adding depth to character strategy and gameplay choices.