Vampires can only feed on living monsters with blood.
I noticed I was able to "feed" on a fog cloud, which is quite strange because it doesn't have blood... Looking at the original SLASH'EM vampire, this seems to be a thing dating back where you can "feed on the life force". This just doesn't make sense to me even by NetHack standards so I'm updating it. The small amount of nutrition vampires might have gotten from these types of monsters should now be made up by the extra potions of blood that generate.