Remove the satiated hunger check from vampire feeding.
Player vampires are *not* smart enough to stop feeding if they might have trouble getting it down. If you feed while satiated, there is a chance you'll overfeed and be occupied/paralyzed for 2 turns. This was how the vampires worked in SLASH'EM, and I think the hunger check was added in UnNetHack, but thematically I think that safety check goes against the spirit of vampires. They are bloodythirsty and are typically pictured as not stopping until they are blood drunk or dead.