Tools have the same recharging limits as wands.
My previous nerf for horns of plenty was a little short sighted - I should have aimed higher. This is a more broad change that applies to any chargeable tools - barring the Bell of Opening and artifact tools like any Orbs of course. It uses the same calculations as wands, so the max number of recharges is 7. After the first recharge there is always a chance that the tool will just crumble from stress.

This is basically a nerf to some infinite combos with the Tourist's QA, but also to generally limit the utility of tools to a reasonable amount.