Shield skill grants 4x more skill than before.
I noticed my shield skill was training exceptionally slow in a valkyrie game. When I had brought in shield skill I neglected to account for the increased skill requirements for training. The skill requirements are roughly 2.5x more than they are in vanilla but since shields train fairly infrequently we'll be a bit more generous.