Move Wintercloak so she's inside the Sokoban zoo.
When she was behind the prize door, Wintercloak would be able to pickup the scroll of scare monster which felt a bit strange.

This also makes the fight harder since she would be frequently be leftover and isolated after the main zoo fight. Now Wintercloak is right behind the door to the zoo forcing a confrontation with all the zoo as backup.