Smarter vampire biting.
This change modifies how the vampire bite attack is handled, addressing two main points:
1) Vampires currently gain an unfair advantage with an additional attack.
2) Vampire feeding mechanics should be more effective and tied to the vampire's hunger status.

On the first point, the extra attack adds an additional 1d6 bite damage plus a potential life drain effect on the victim. This is akin to granting the vampire an extra quarterstaff that drains life on every attack round. Since its inception in SLASH'EM, this mechanic has been overpowered, making non-fighter roles like vampire archaeologists comparable to — or even better than — valkyries. The life drain component alone is sufficient to defeat most low-level monsters in just a few turns, whilst the double attack can eliminate most monsters in the early game with minimal effort.

To address this, my solution is to limit the vampire's attack options to either biting or attacking with a weapon until the player reaches a higher level. If the vampire is hungry (or close to hungry), they will always bite and attempt to feed first. Otherwise, there will be a 50/50 chance of either biting or attacking with a normal weapon. Once the player reaches level 15, the ability to perform both actions in the same round will unlock.

Furthermore, if a player wishes to only use the weapon, they can always bypass the bite attack by forcefighting (via "F"). This is a nice feature from EvilHack where there are illithids and other races that might not want to use their innate attacks on something dangerous.