Fix monsters reading scrolls of flood.
This addresses a strange bug that occurs deep in the bowels of do_clear_area somewhere in view_from. I have looked at this problem for a while and from different angles, but I cannot figure it out. I think the easy way out is just making this scroll always center on the player, even when read by monsters. This cures the bug and it makes sense for monsters so we're rolling with it.