Refactor vampire biting out to it's own special attack AD_VAMP (from dNetHack).
This has been a long time coming, but it's finally time. The vampire attack differs from drain life in enough ways to make the drain life code ugly and annoying to work with. This cleans up that problem and strikes a clear separation between the two.

I didn't really follow the dnh implementation, just used it for inspiration.

Some interesting things to note:
* I removed the requirement for the monster to have blood. The original SLASH'EM mechanic has vampires feeding "on lifeblood", which doesn't actually require the monster to have blood.
* Feeding will heal the player and monsters, making vampires harder to fight.