Fine tune frequency of grung poison attack and passives.
I think it's fair to say that their poison cannot trigger on every attack 100% of the time, that is too good. I added a mon MC check to attacking them bare-handed and also cut it to 1 in 3. Also lowered the base bonus to 1d6 with a 10% chance of instakill.

Also adjusted the frequencies for mhitm and mhitu.